requirements.txt
streamlit
pandas
Pillow
cv2
requests
json